% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarConversion.R
\name{julianDay2mjd}
\alias{julianDay2mjd}
\title{Convert Julian Day to Modified Julian Day}
\usage{
julianDay2mjd(jd)
}
\arguments{
\item{jd}{Modified Julian Day}
}
\value{
Corresponding Modified Julian Day number
}
\description{
Convert Julian Day to Modified Julian Day
}
\examples{
julianDay2mjd(0)

}
\seealso{
mjd2julianDay
}
