% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarConversion.R
\name{mjd2julianDay}
\alias{mjd2julianDay}
\title{Convert Modified Julian Day to Julian Day}
\usage{
mjd2julianDay(mjd)
}
\arguments{
\item{mjd}{Modified Julian Day}
}
\value{
Corresponding Julian Day number
}
\description{
Converts a Modified Julian Day (MJD) number (as used by some satellite
systems) to a Julian Day number. A MJD of zero corresponds to midnight
on November 17, 1858.
}
\examples{
mjd2julianDay(0)

}
\seealso{
julianDay2mjd
}
