% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnalYear.R
\name{regnalYear}
\alias{regnalYear}
\title{Convert date to regnal year}
\usage{
regnalYear(date, region = "UK", system = "UK", abbr = FALSE)
}
\arguments{
\item{date}{Date to convert (YYYY-mm-dd)}

\item{region}{Monarchy to use (currently UK)}

\item{system}{Date system of input (default UK)}

\item{abbr}{Abbreviate monarch names}
}
\value{
Regnal year and reigning monarch
}
\description{
Convert date to regnal year
}
\examples{
regnalYear("1067-01-01")
regnalYear("1067-01-01", region="UK", system="Julian")
regnalYear("1067-01-01", abbr=TRUE)

}
