% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotland_regnal.R
\docType{data}
\name{ScotlandRegnal}
\alias{ScotlandRegnal}
\title{Start dates of Scottish regnal years.}
\format{
A data frame with 4 variables:
\describe{
  \item{Reign}{Monarch, full name}
  \item{JulianStart}{Start date in Julian system, YYYY-MM-DD}
  \item{GregorianStart}{Start date in Gregorian system, YYYY-MM-DD}
  \item{Abbreviation}{Abbreviated form of data in Reign}
  \item{RegnalYear}{Used to store compound and complex years}
}
}
\source{
Dunbar, AH (1906) A Revised Chronology of Scottish History 1005 - 1625.
}
\usage{
ScotlandRegnal
}
\description{
A dataset containing the start date of regnal years, the ruling
monarch(s) of those regnal years, and abbreviated name forms of the
monarch(s).
}
\keyword{datasets}
