% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_regnal.R
\docType{data}
\name{UKregnal}
\alias{UKregnal}
\title{Start dates of UK regnal years.}
\format{
A data frame with 4 variables:
\describe{
  \item{Reign}{Monarch, full name}
  \item{JulianStart}{Start date in Julian system, YYYY-MM-DD}
  \item{Abbreviation}{Abbreviated form of data in Reign}
  \item{Notes}{Miscellaneous related data}
}
}
\source{
Harvey, P (1975) The Oxford Companion to English Literature.
}
\usage{
UKregnal
}
\description{
A dataset containing the start date of regnal years, the ruling
monarch(s) of those regnal years, and abbreviated name forms of the
monarch(s).
}
\keyword{datasets}
